Expose `parse` through public interface
