+++ cypher-codemirror/src/codemirror-cypher.js
@@ -19,7 +19,7 @@
  */
 
 import codemirror from 'codemirror';
-import { CypherEditorSupport, TreeUtils } from 'cypher-editor-support/src/index';
+import { CypherEditorSupport, TreeUtils, parse as importedParse } from 'cypher-editor-support/src/index';
 import './codemirror-cypher-mode';
 
 function translatePosition(from, to) {
@@ -113,3 +113,5 @@ export function createCypherEditor(input, settings) {
 
   return { editor, editorSupport };
 }
+
+export const parse = importedParse;
